requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
yoomoney==0.3.0
requests==2.31.0
aiosqlite==0.19.0
aiocryptopay==0.6.0